Extract EventWriter
